+++ src/driver.rs
@@ -2,7 +2,7 @@
 //! different databases exist elsewhere.
 
 use error::Error;
-use document::Document;
+use values::Value;
 use structure::Structure;
 use request::Request;
 
@@ -13,11 +13,11 @@ pub trait Driver {
   fn get_structure() -> Result<Structure, Error>;
 
   /// Send a request to the database.
-  fn request(request: Request) -> Result<Vec<Document>, Error>;
+  fn request(request: Request) -> Result<Vec<Value>, Error>;
 
   /// Send a request to the database and guarantee it only ever effects a
   /// single document.
-  fn request_one(request: Request) -> Result<Document, Error> {
+  fn request_one(request: Request) -> Result<Value, Error> {
     // TODO: default implementation which verifies that the request is only
     // affecting a single thing. This may be done by checking if the collection
     // primary key is specified in a filter.
@@ -27,7 +27,7 @@ pub trait Driver {
   /// Send multiple requests at once. If one request fails, **all other
   /// requests must also fail**. If a driver author wishes they may also
   /// optimize these requests.
-  fn requests(requests: Vec<Request>) -> Result<Vec<Vec<Document>>, Error> {
+  fn requests(requests: Vec<Request>) -> Result<Vec<Vec<Value>>, Error> {
     // TODO: default implementation which just runs all the requests and returns
     // a result.
     unimplemented!()
