+++ src/range.rs
@@ -1,50 +1,4 @@
-//! Defines types used in trimming and transforming a set of documents in a
-//! database.
-
-use document::Property;
-use document::Value;
-
-/// A recursive filter condition.
-pub enum Filter {
-  /// Combine multiple filters with an “and” operator.
-  And(Vec<Box<Filter>>),
-
-  /// Combine multiple filters with an “or” operator.
-  Or(Vec<Box<Filter>>),
-
-  /// Inverts the filter.
-  Not(Box<Filter>),
-
-  /// The basic condition of a filter.
-  Condition {
-    /// The property to use in evaluating the condition.
-    property: Property,
-
-    /// The condition to be used in evaluating.
-    condition: FilterCondition
-  }
-}
-
-pub enum FilterCondition {
-  Equal(Value),
-  OneOf(Vec<Value>),
-  GreaterThan(Value),
-  LessThan(Value)
-}
-
-/// A single way in which to order a collection of documents.
-pub struct Ordering {
-  /// The property to order against.
-  property: Property,
-
-  /// The direction to order in. Either ascending or descending.
-  direction: OrderDirection
-}
-
-pub enum OrderDirection {
-  Ascending,
-  Descending
-}
+// TODO: Make this it's own module or use another implementation.
 
 /// A numeric range. Inspired by Haskell‘s [ranged-sets][1].
 ///
