+++ src/request.rs
@@ -2,10 +2,8 @@
 //! `Request` enum contains all such requests. This module only contains the
 //! `Request` enum.
 
-use patch::*;
-use selection::*;
+use values::*;
 use structure::Collection;
-use document::*;
 
 /// Any driver request. All requests will return an array of documents.
 pub enum Request {
@@ -18,12 +16,10 @@ pub enum Request {
   Create {
     /// The collection to create the documents in.
     collection: Collection,
-
     /// The documents to be created in the database.
     documents: Vec<Document>,
-
     /// The properties to be returned of the created documents.
-    returning: ReturnSet,
+    returning: ReturnSet
   },
 
   /// The read request. Designed after a [SQL `SELECT` statement][1]. Also, a
@@ -39,16 +35,12 @@ pub enum Request {
   Read {
     /// The collection of which to read.
     collection: Collection,
-
     /// The filter with which to narrow the documents returned.
     filter: Filter,
-
     /// The order in which to return the documents.
     order: Vec<Ordering>,
-
     /// A specific range of documents to be read.
     range: Range,
-
     /// The properties to be returned of the read documents.
     returning: ReturnSet
   },
@@ -62,13 +54,10 @@ pub enum Request {
   Update {
     /// The collection in which to update documents.
     collection: Collection,
-
     /// A filter limiting the number of documents to be updated.
     filter: Filter,
-
     /// The patches to be applied to the set of documents.
     patches: Vec<Patch>,
-
     /// The properties to be returned of the updated documents.
     returning: ReturnSet
   },
@@ -82,10 +71,8 @@ pub enum Request {
   Delete {
     /// The collection to delete documents from.
     collection: Collection,
-
     /// A filter limiting the number of documents deleted.
     filter: Filter,
-
     /// What properties to be returned of the deleted documents.
     returning: ReturnSet
   }
@@ -95,7 +82,6 @@ pub enum Request {
 pub enum ReturnSet {
   /// All of the properties in the document.
   All,
-
   /// Only some of the document‘s properties.
   Some(Vec<Property>)
 }
